+++ noodles-bcf/src/reader/query.rs
@@ -6,8 +6,9 @@ use std::{
 use noodles_bgzf as bgzf;
 use noodles_core::{region::Interval, Position};
 use noodles_csi::index::reference_sequence::bin::Chunk;
+use noodles_vcf as vcf;
 
-use crate::lazy;
+use crate::header::StringMaps;
 
 use super::Reader;
 
@@ -20,27 +21,29 @@ enum State {
 /// An iterator over records of a BCF reader that intersects a given region.
 ///
 /// This is created by calling [`Reader::query`].
-pub struct Query<'a, R>
+pub struct Query<'r, 'h, R>
 where
     R: Read + Seek,
 {
-    reader: &'a mut Reader<bgzf::Reader<R>>,
+    reader: &'r mut Reader<bgzf::Reader<R>>,
 
+    header: &'h vcf::Header,
     chunks: vec::IntoIter<Chunk>,
 
     chromosome_id: usize,
     interval: Interval,
 
     state: State,
-    record: lazy::Record,
+    record: vcf::Record,
 }
 
-impl<'a, R> Query<'a, R>
+impl<'r, 'h, R> Query<'r, 'h, R>
 where
     R: Read + Seek,
 {
     pub(crate) fn new(
-        reader: &'a mut Reader<bgzf::Reader<R>>,
+        reader: &'r mut Reader<bgzf::Reader<R>>,
+        header: &'h vcf::Header,
         chunks: Vec<Chunk>,
         chromosome_id: usize,
         interval: Interval,
@@ -48,19 +51,20 @@ where
         Self {
             reader,
 
+            header,
             chunks: chunks.into_iter(),
 
             chromosome_id,
             interval,
 
             state: State::Seek,
-            record: lazy::Record::default(),
+            record: vcf::Record::default(),
         }
     }
 
-    fn read_record(&mut self) -> io::Result<Option<lazy::Record>> {
+    fn read_record(&mut self) -> io::Result<Option<vcf::Record>> {
         self.reader
-            .read_lazy_record(&mut self.record)
+            .read_record(self.header, &mut self.record)
             .map(|n| match n {
                 0 => None,
                 _ => Some(self.record.clone()),
@@ -68,11 +72,11 @@ where
     }
 }
 
-impl<'a, R> Iterator for Query<'a, R>
+impl<'r, 'h, R> Iterator for Query<'r, 'h, R>
 where
     R: Read + Seek,
 {
-    type Item = io::Result<lazy::Record>;
+    type Item = io::Result<vcf::Record>;
 
     fn next(&mut self) -> Option<Self::Item> {
         loop {
@@ -95,7 +99,12 @@ where
                             self.state = State::Seek;
                         }
 
-                        match intersects(&record, self.chromosome_id, self.interval) {
+                        match intersects(
+                            self.reader.string_maps(),
+                            &record,
+                            self.chromosome_id,
+                            self.interval,
+                        ) {
                             Ok(true) => return Some(Ok(record)),
                             Ok(false) => {}
                             Err(e) => return Some(Err(e)),
@@ -110,19 +119,34 @@ where
     }
 }
 
-pub(crate) fn intersects(
-    record: &lazy::Record,
+fn intersects(
+    string_maps: &StringMaps,
+    record: &vcf::Record,
     chromosome_id: usize,
     region_interval: Interval,
 ) -> io::Result<bool> {
-    let id = record.chromosome_id();
+    let chromosome = record.chromosome();
+
+    let id = string_maps
+        .contigs()
+        .get_index_of(&chromosome.to_string())
+        .ok_or_else(|| {
+            io::Error::new(
+                io::ErrorKind::InvalidInput,
+                format!("chromosome does not exist in contigs: {chromosome}"),
+            )
+        })?;
 
     let start = Position::try_from(usize::from(record.position()))
         .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
 
-    let end = record.end().map(usize::from).and_then(|n| {
-        Position::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-    })?;
+    let end = record
+        .end()
+        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+        .map(usize::from)
+        .and_then(|n| {
+            Position::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+        })?;
 
     let record_interval = Interval::from(start..=end);